bert_tf2/requirement.txt
tensorflow==2.2.0-rc1
bert-for-tf2
sentencepiece
pandas
matplotlib
seaborn
sklearn
tqdm
